Rename PandasTrackDatasetFactory class implementation to avoid naming conflicts
